requirements.txt
confluent_kafka==2.2.0
fastapi==0.104.0
pydantic==2.4.2
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.22
streamlit==1.27.2
stripe==7.0.0
uvicorn==0.23.2
